Issue-201/624: Updated SBT. Moved around a few classes. Piped in new test target into makefile.
