--- src/main/scala/org/openhorizon/exchangeapi/route/administration/GetAdminOrgStatusResponse.scala
@@ -1,3 +1,0 @@
-package org.openhorizon.exchangeapi.route.administration
-
-final case class GetAdminOrgStatusResponse(msg: String, nodes: Map[String, Int])
